# Ensure range index incremental for data source op like `md.read_csv`