--- a/mars/dataframe/datasource/read_csv.py
+++ b/mars/dataframe/datasource/read_csv.py
@@ -35,7 +35,8 @@ from ..arrays import ArrowStringDtype
 from ..core import IndexValue
 from ..utils import parse_index, build_empty_df, standardize_range_index, \
     to_arrow_dtypes, contain_arrow_dtype
-from .core import HeadOptimizedDataSource, ColumnPruneSupportedDataSourceMixin
+from .core import IncrementalIndexDatasource, ColumnPruneSupportedDataSourceMixin, \
+    IncrementalIndexDataSourceMixin
 
 
 cudf = lazy_import('cudf', globals=globals())
@@ -82,7 +83,9 @@ def _find_chunk_start_end(f, offset, size):
     return start, end
 
 
-class DataFrameReadCSV(HeadOptimizedDataSource, ColumnPruneSupportedDataSourceMixin):
+class DataFrameReadCSV(IncrementalIndexDatasource,
+                       ColumnPruneSupportedDataSourceMixin,
+                       IncrementalIndexDataSourceMixin):
     _op_type_ = OperandDef.READ_CSV
 
     _path = AnyField('path')